+++ packages/react/src/components/InitialScreen/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The `InitialScreen` component provides a consistent layout and spacing for displaying a landing or welcome screen. It introduces users to an application or service, offering options to access more information or proceed further.
 
+## 2025-10-16 `1.0.4`
+
+### 💡 Others
+
+- Fixes background color of `InitialScreen.Card`. ([#111](https://github.com/TiendaNube/nimbus-patterns/pull/111) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-04-24 `1.0.3`
 
 ### 💡 Others